Removed accidental adding of output parser
